Added Webinar in Progress section  under Webinar Completed.
Webinar in progress title was added under webinar in progress blade file, change the variable in relation to in progress.

The controller panel changed the value of the previous code line to null, and removed code line that was for date completed and e- certificate.
The variables from the previous code was changed into progresswebinars.

Co-Authored-By: Kenneth Lorenzo D. Tiongson <91168024+kenzo1011@users.noreply.github.com>